+++ src/main/java/com/example/moviewbackend/MoviewBackendApplication.java
@@ -9,6 +9,7 @@ import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 public class MoviewBackendApplication {
 
     public static void main(String[] args) {
+
         SpringApplication.run(MoviewBackendApplication.class, args);
     }
