+++ src/main/java/com/example/moviewbackend/dto/SignupRequestDto.java
@@ -3,15 +3,17 @@ package com.example.moviewbackend.dto;
 import com.example.moviewbackend.entity.UserRoleEnum;
 import jakarta.validation.constraints.Pattern;
 import lombok.Getter;
+import lombok.Setter;
 
 @Getter
+@Setter
 public class SignupRequestDto {
     @Pattern(regexp = "^[a-z0-9]{4,10}$",
             message = "최소 4자 이상, 10자 이하이며 알파벳 소문자(a~z), 숫자(0~9) 로 구성되어야 합니다.")
     private String nickname;
 
     //@Email
-    @Pattern(regexp = "")
+    @Pattern(regexp = "^[a-zA-Z0-9+-\\_.]+@[a-zA-Z0-9-]+\\.[a-zA-Z0-9-.]+$")
     private String email;
 
     @Pattern(regexp = "^[a-zA-Z0-9!@#$%^&*()_+{}:\"<>?,.\\\\/]{8,15}$",
