+++ src/main/java/com/example/moviewbackend/jwt/JwtAuthorizationFilter.java
@@ -7,11 +7,13 @@ import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContext;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.util.StringUtils;
 import org.springframework.web.filter.OncePerRequestFilter;
 
@@ -47,10 +49,23 @@ public class JwtAuthorizationFilter extends OncePerRequestFilter {
 
             try {
                 setAuthentication(info.getSubject());
+            } catch (UsernameNotFoundException e) {
+                log.error("User not found: {}", e.getMessage());
+
+                res.setStatus(HttpServletResponse.SC_NOT_FOUND);
+                return;
+            } catch (BadCredentialsException e) {
+                log.error("Invalid credentials: {}", e.getMessage());
+
+                res.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
+                return;
             } catch (Exception e) {
-                log.error(e.getMessage());
+                log.error("Error occurred during authentication: {}", e.getMessage());
+
+                res.setStatus(HttpServletResponse.SC_INTERNAL_SERVER_ERROR);
                 return;
             }
+
         }
 
         filterChain.doFilter(req, res);
