+++ src/main/java/com/example/moviewbackend/repository/UserRepository.java
@@ -8,6 +8,5 @@ import java.util.Optional;
 public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByEmail(String email);
     Optional<User> findByNickname(String nickname);
-
     Optional<User> findByKakaoId(Long kakaoId);
 }
