+++ src/main/java/com/example/moviewbackend/security/UserDetailsImpl.java
@@ -28,13 +28,14 @@ public class UserDetailsImpl implements UserDetails {
 
     @Override
     public String getUsername() {
-        return user.getNickname();
+        return user.getEmail();
     }
 
     @Override
     public Collection<? extends GrantedAuthority> getAuthorities() {
         UserRoleEnum role = user.getRole();
-        String authority = role.getAuthority();
+        String authority = (role != null) ? role.getAuthority() : "DEFAULT_ROLE";
+
 
         SimpleGrantedAuthority simpleGrantedAuthority = new SimpleGrantedAuthority(authority);
         Collection<GrantedAuthority> authorities = new ArrayList<>();
