+++ src/main/java/com/example/moviewbackend/service/UserService.java
@@ -1,16 +1,12 @@
 package com.example.moviewbackend.service;
 
-import com.example.moviewbackend.dto.KakaoUserInfoDto;
 import com.example.moviewbackend.dto.SignupRequestDto;
 import com.example.moviewbackend.entity.User;
-import com.example.moviewbackend.entity.UserRoleEnum;
 import com.example.moviewbackend.jwt.JwtUtil;
 import com.example.moviewbackend.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
-
-import java.util.UUID;
 
 @Service
 @RequiredArgsConstructor
@@ -35,7 +31,7 @@ public class UserService {
             throw new IllegalArgumentException("이미 존재하는 닉네임입니다.");
         }
 
-        User user = new User(email, password);
+        User user = new User(email, password, nickname);
         userRepository.save(user);
     }
 
@@ -47,33 +43,4 @@ public class UserService {
 
     }
 
-    private User registerKakaoUserIfNeeded(KakaoUserInfoDto kakaoUserInfo) {
-        // DB 에 중복된 Kakao Id 가 있는지 확인
-        Long kakaoId = kakaoUserInfo.getId();
-        User kakaoUser = userRepository.findByKakaoId(kakaoId).orElse(null);
-
-        if (kakaoUser == null) {
-            // 카카오 사용자 email 동일한 email 가진 회원이 있는지 확인
-            String kakaoEmail = kakaoUserInfo.getEmail();
-            User sameEmailUser = userRepository.findByEmail(kakaoEmail).orElse(null);
-            if (sameEmailUser != null) {
-                kakaoUser = sameEmailUser;
-                // 기존 회원정보에 카카오 Id 추가
-                kakaoUser = kakaoUser.kakaoIdUpdate(kakaoId);
-            } else {
-                // 신규 회원가입
-                // password: random UUID
-                String password = UUID.randomUUID().toString();
-                String encodedPassword = passwordEncoder.encode(password);
-
-                // email: kakao email
-                String email = kakaoUserInfo.getEmail();
-
-                kakaoUser = new User(kakaoUserInfo.getNickname(), encodedPassword, email, UserRoleEnum.USER, kakaoId);
-            }
-
-            userRepository.save(kakaoUser);
-        }
-        return kakaoUser;
-    }
 }
